// src/app/models/complaint.enum.ts
export enum COMPLAINTENUM {
    ABDOMINALPAIN = 'abdominal pain',
    ALLERGY = 'allergy',
    BODYMALAISE = 'body malaise',
    CHESTPAIN = 'chest pain',
    COLD = 'cold',
    DYSMENORRHEA = 'dysmenorrhea',
    HEADACHE = 'headache',
    NAUSEA = 'nausea',
    SKIN_RASH = 'skin rash',
    SPRAIN = 'sprain',
    VOMNITING = 'vomiting',
    WOUND = 'wound',
    OTHERS = 'others'
}